Moved MyFakeClient from main.go to retester_test.go
